grasp: moved env files to new ./envs folder
