+++ aegis_gym/rsl/grasp_train.py
@@ -7,7 +7,7 @@ from rsl_rl.runners import OnPolicyRunner
 
 from behavior_cloning import BehaviorCloning
 from utils import check_rsl_rl_version, load_teacher_policy
-from grasp_env import GraspEnv
+from aegis_gym.aegis_gym.rsl.envs.grasp_env import GraspEnv
 from grasp_cfgs import get_task_cfgs, get_rl_cfg, get_bc_cfg
